+++ packages/atomic/src/components/facets/atomic-segmented-facet-scrollable/atomic-segmented-facet-scrollable.tsx
@@ -45,7 +45,7 @@ export class AtomicSegmentedFacetScrollable implements InitializableComponent {
 
   private horizontalScroll?: HTMLDivElement;
   @State() private hideLeftArrow = true;
-  @State() private hideRightArrow = false;
+  @State() private hideRightArrow = true;
   private observer!: ResizeObserver;
 
   public initialize() {
@@ -65,6 +65,7 @@ export class AtomicSegmentedFacetScrollable implements InitializableComponent {
 
   @Listen('wheel')
   @Listen('touchmove')
+  @Listen('keydown')
   handleScroll() {
     if (!this.horizontalScroll) {
       return;
@@ -107,20 +108,36 @@ export class AtomicSegmentedFacetScrollable implements InitializableComponent {
 
   private renderArrowClass(direction: ArrowDirection) {
     if (direction === 'left') {
-      return 'left-0 ' + (this.hideLeftArrow ? 'invisible opacity-0' : '');
+      return (
+        'left-0 ' +
+        (this.hideLeftArrow || !this.searchStatusState.firstSearchExecuted
+          ? 'invisible opacity-0'
+          : '')
+      );
     } else {
-      return 'right-0 ' + (this.hideRightArrow ? 'invisible opacity-0' : '');
+      return (
+        'right-0 ' +
+        (this.hideRightArrow || !this.searchStatusState.firstSearchExecuted
+          ? 'invisible opacity-0'
+          : '')
+      );
     }
   }
 
   private renderFadeClass(direction: ArrowDirection) {
     if (direction === 'left') {
       return (
-        'bg-gradient-to-r left-0 ' + (this.hideLeftArrow ? 'opacity-0' : '')
+        'bg-gradient-to-r left-0 ' +
+        (this.hideLeftArrow || !this.searchStatusState.firstSearchExecuted
+          ? 'opacity-0'
+          : '')
       );
     } else {
       return (
-        'bg-gradient-to-l right-0 ' + (this.hideRightArrow ? 'opacity-0' : '')
+        'bg-gradient-to-l right-0 ' +
+        (this.hideRightArrow || !this.searchStatusState.firstSearchExecuted
+          ? 'opacity-0'
+          : '')
       );
     }
   }