+++ packages/atomic/src/components/search/atomic-search-interface/atomic-search-interface.tsx
@@ -352,8 +352,8 @@ export class AtomicSearchInterface
 
     if (options.search) {
       return {
-        ...options.search,
         ...searchConfigFromProps,
+        ...options.search,
       };
     }