+++ packages/headless/src/controllers/index.ts
@@ -194,6 +194,8 @@ export type {
 export type {
   InstantResults,
   InstantResultsState,
+  InstantResultProps,
+  InstantResultOptions,
 } from './instant-results/instant-results';
 export {buildInstantResults} from './instant-results/instant-results';
 
@@ -229,6 +231,8 @@ export type {
   Tab,
 } from './tab/headless-tab';
 export {buildTab} from './tab/headless-tab';
+
+export type {CoreFacetManager} from './core/facet-manager/headless-core-facet-manager';
 
 export type {
   FacetManagerPayload,