+++ packages/headless/src/controllers/insight/breadcrumb-manager/headless-insight-breadcrumb-manager.ts
@@ -1,0 +1,247 @@
+import {InsightEngine} from '../../../app/insight-engine/insight-engine';
+import {
+  categoryFacetSet,
+  configuration,
+  dateFacetSet,
+  facetSet,
+  numericFacetSet,
+  search,
+} from '../../../app/reducers';
+import {deselectAllCategoryFacetValues} from '../../../features/facets/category-facet-set/category-facet-set-actions';
+import {categoryFacetSelectedValuesSelector} from '../../../features/facets/category-facet-set/category-facet-set-selectors';
+import {
+  toggleSelectFacetValue,
+  updateFreezeCurrentValues,
+} from '../../../features/facets/facet-set/facet-set-actions';
+import {facetResponseSelectedValuesSelector} from '../../../features/facets/facet-set/facet-set-selectors';
+import {FacetValue} from '../../../features/facets/facet-set/interfaces/response';
+import {toggleSelectDateFacetValue} from '../../../features/facets/range-facets/date-facet-set/date-facet-actions';
+import {dateFacetSelectedValuesSelector} from '../../../features/facets/range-facets/date-facet-set/date-facet-selectors';
+import {DateFacetValue} from '../../../features/facets/range-facets/date-facet-set/interfaces/response';
+import {NumericFacetValue} from '../../../features/facets/range-facets/numeric-facet-set/interfaces/response';
+import {toggleSelectNumericFacetValue} from '../../../features/facets/range-facets/numeric-facet-set/numeric-facet-actions';
+import {numericFacetSelectedValuesSelector} from '../../../features/facets/range-facets/numeric-facet-set/numeric-facet-selectors';
+import {
+  logInsightStaticFilterDeselect,
+  toggleSelectStaticFilterValue,
+} from '../../../features/static-filter-set/static-filter-set-actions';
+import {
+  StaticFilterSlice,
+  StaticFilterValue,
+} from '../../../features/static-filter-set/static-filter-set-state';
+import {
+  CategoryFacetSection,
+  ConfigurationSection,
+  DateFacetSection,
+  FacetSection,
+  NumericFacetSection,
+  SearchSection,
+} from '../../../state/state-sections';
+import {
+  Breadcrumb,
+  BreadcrumbManager,
+  BreadcrumbManagerState,
+  BreadcrumbValue,
+  buildCoreBreadcrumbManager,
+  CategoryFacetBreadcrumb,
+  DateFacetBreadcrumb,
+  DeselectableValue,
+  FacetBreadcrumb,
+  getBreadcrumbs,
+  NumericFacetBreadcrumb,
+  StaticFilterBreadcrumb,
+} from '../../core/breadcrumb-manager/headless-core-breadcrumb-manager';
+import {executeSearch} from '../../../features/insight-search/insight-search-actions';
+import {logNumericFacetBreadcrumb} from '../../../features/facets/range-facets/numeric-facet-set/numeric-facet-insight-analytics-actions';
+import {logDateFacetBreadcrumb} from '../../../features/facets/range-facets/date-facet-set/date-facet-insight-analytics-actions';
+import {logClearBreadcrumbs} from '../../../features/facets/generic/facet-generic-insight-analytics-actions';
+import {loadReducerError} from '../../../utils/errors';
+import {logCategoryFacetBreadcrumb} from '../../../features/facets/category-facet-set/category-facet-set-insight-analytics-actions';
+import {logFacetBreadcrumb} from '../../../features/facets/facet-set/facet-set-insight-analytics-actions';
+
+export type {
+  NumericFacetBreadcrumb,
+  FacetBreadcrumb,
+  DateFacetBreadcrumb,
+  CategoryFacetBreadcrumb,
+  StaticFilterBreadcrumb,
+  Breadcrumb,
+  BreadcrumbValue,
+  BreadcrumbManagerState,
+  BreadcrumbManager,
+  DeselectableValue,
+};
+/**
+ * Creates an insight `BreadcrumbManager` controller instance.
+ *
+ * @param engine - The headless engine.
+ * @returns A `BreadcrumbManager` controller instance.
+ */
+export function buildBreadcrumbManager(
+  engine: InsightEngine
+): BreadcrumbManager {
+  if (!loadBreadcrumbManagerReducers(engine)) {
+    throw loadReducerError;
+  }
+  const controller = buildCoreBreadcrumbManager(engine);
+  const {dispatch} = engine;
+  const getState = () => engine.state;
+
+  const getFacetBreadcrumbs = (): FacetBreadcrumb[] => {
+    return getBreadcrumbs<FacetValue>(
+      engine,
+      getState().facetSet,
+      ({facetId, selection}) => {
+        const analyticsAction = logFacetBreadcrumb({
+          facetId: facetId,
+          facetValue: selection.value,
+        });
+        dispatch(toggleSelectFacetValue({facetId, selection}));
+        dispatch(
+          updateFreezeCurrentValues({facetId, freezeCurrentValues: false})
+        );
+        dispatch(executeSearch(analyticsAction));
+      },
+      facetResponseSelectedValuesSelector
+    );
+  };
+
+  const getNumericFacetBreadcrumbs = (): NumericFacetBreadcrumb[] => {
+    return getBreadcrumbs<NumericFacetValue>(
+      engine,
+      getState().numericFacetSet,
+      (payload) => {
+        dispatch(toggleSelectNumericFacetValue(payload));
+        dispatch(executeSearch(logNumericFacetBreadcrumb(payload)));
+      },
+      numericFacetSelectedValuesSelector
+    );
+  };
+
+  const getDateFacetBreadcrumbs = (): DateFacetBreadcrumb[] => {
+    return getBreadcrumbs<DateFacetValue>(
+      engine,
+      getState().dateFacetSet,
+      (payload) => {
+        dispatch(toggleSelectDateFacetValue(payload));
+        dispatch(executeSearch(logDateFacetBreadcrumb(payload)));
+      },
+      dateFacetSelectedValuesSelector
+    );
+  };
+
+  const buildCategoryFacetBreadcrumb = (facetId: string) => {
+    const path = categoryFacetSelectedValuesSelector(getState(), facetId);
+    return {
+      facetId,
+      field: getState().categoryFacetSet[facetId]!.request.field,
+      path,
+      deselect: () => {
+        dispatch(deselectAllCategoryFacetValues(facetId));
+        dispatch(
+          executeSearch(
+            logCategoryFacetBreadcrumb({
+              categoryFacetPath: path.map(
+                (categoryFacetValue) => categoryFacetValue.value
+              ),
+              categoryFacetId: facetId,
+            })
+          )
+        );
+      },
+    };
+  };
+
+  const getCategoryFacetBreadcrumbs = (): CategoryFacetBreadcrumb[] => {
+    return Object.keys(getState().categoryFacetSet)
+      .map(buildCategoryFacetBreadcrumb)
+      .filter((breadcrumb) => breadcrumb.path.length);
+  };
+
+  const getStaticFilterBreadcrumbs = (): StaticFilterBreadcrumb[] => {
+    const set = getState().staticFilterSet || {};
+    return Object.values(set).map(buildStaticFilterBreadcrumb);
+  };
+
+  const buildStaticFilterBreadcrumb = (
+    filter: StaticFilterSlice
+  ): StaticFilterBreadcrumb => {
+    const {id, values: filterValues} = filter;
+    const values = filterValues
+      .filter((value) => value.state === 'selected')
+      .map((value) => buildStaticFilterBreadcrumbValue(id, value));
+
+    return {id, values};
+  };
+
+  const buildStaticFilterBreadcrumbValue = (
+    id: string,
+    value: StaticFilterValue
+  ) => {
+    return {
+      value,
+      deselect: () => {
+        const {caption, expression} = value;
+        const analytics = logInsightStaticFilterDeselect({
+          staticFilterId: id,
+          staticFilterValue: {caption, expression},
+        });
+
+        dispatch(toggleSelectStaticFilterValue({id, value}));
+        dispatch(executeSearch(analytics));
+      },
+    };
+  };
+
+  function hasBreadcrumbs() {
+    return !![
+      ...getFacetBreadcrumbs(),
+      ...getNumericFacetBreadcrumbs(),
+      ...getDateFacetBreadcrumbs(),
+      ...getCategoryFacetBreadcrumbs(),
+      ...getStaticFilterBreadcrumbs(),
+    ].length;
+  }
+
+  return {
+    ...controller,
+
+    get state() {
+      return {
+        facetBreadcrumbs: getFacetBreadcrumbs(),
+        categoryFacetBreadcrumbs: getCategoryFacetBreadcrumbs(),
+        numericFacetBreadcrumbs: getNumericFacetBreadcrumbs(),
+        dateFacetBreadcrumbs: getDateFacetBreadcrumbs(),
+        staticFilterBreadcrumbs: getStaticFilterBreadcrumbs(),
+        hasBreadcrumbs: hasBreadcrumbs(),
+      };
+    },
+
+    deselectAll: () => {
+      controller.deselectAll();
+      dispatch(executeSearch(logClearBreadcrumbs()));
+    },
+  };
+}
+
+function loadBreadcrumbManagerReducers(
+  engine: InsightEngine
+): engine is InsightEngine<
+  ConfigurationSection &
+    SearchSection &
+    FacetSection &
+    NumericFacetSection &
+    DateFacetSection &
+    CategoryFacetSection
+> {
+  engine.addReducers({
+    configuration,
+    search,
+    facetSet,
+    numericFacetSet,
+    dateFacetSet,
+    categoryFacetSet,
+  });
+
+  return true;
+}