+++ packages/headless/src/features/facets/generic/facet-generic-insight-analytics-actions.ts
@@ -1,0 +1,13 @@
+import {
+  AnalyticsType,
+  makeInsightAnalyticsAction,
+} from '../../analytics/analytics-utils';
+
+export const logClearBreadcrumbs = () =>
+  makeInsightAnalyticsAction(
+    'analytics/facet/deselectAllBreadcrumbs',
+    AnalyticsType.Search,
+    (client) => {
+      return client.logBreadcrumbResetAll();
+    }
+  )();