+++ packages/headless/src/insight.index.ts
@@ -39,9 +39,37 @@ export type {
 } from './controllers/controller/headless-controller';
 export {buildController} from './controllers/controller/headless-controller';
 
-export type {InsightInterfaceState} from './features/insight-interface/insight-interface-state';
-export type {InsightInterface} from './controllers/insight-interface/insight-interface';
-export {buildInsightInterface} from './controllers/insight-interface/insight-interface';
+export type {
+  DidYouMean,
+  DidYouMeanState,
+  QueryCorrection,
+  WordCorrection,
+} from './controllers/insight/did-you-mean/headless-insight-did-you-mean';
+export {buildDidYouMean} from './controllers/insight/did-you-mean/headless-insight-did-you-mean';
+
+export type {
+  NumericFacetBreadcrumb,
+  FacetBreadcrumb,
+  DateFacetBreadcrumb,
+  CategoryFacetBreadcrumb,
+  StaticFilterBreadcrumb,
+  Breadcrumb,
+  BreadcrumbValue,
+  BreadcrumbManagerState,
+  BreadcrumbManager,
+  DeselectableValue,
+} from './controllers/insight/breadcrumb-manager/headless-insight-breadcrumb-manager';
+export {buildBreadcrumbManager} from './controllers/insight/breadcrumb-manager/headless-insight-breadcrumb-manager';
+
+export type {
+  SearchParameterManagerProps,
+  SearchParameterManagerInitialState,
+  SearchParameterManagerState,
+  SearchParameterManager,
+  SearchParameters,
+} from './controllers/insight/search-parameter-manager/headless-insight-search-parameter-manager';
+export {buildSearchParameterManager} from './controllers/insight/search-parameter-manager/headless-insight-search-parameter-manager';
+
 export type {
   InsightSearchBox,
   SearchBoxState,
@@ -56,3 +84,8 @@ export type {
 export {buildInsightResultList} from './controllers/insight/result-list/headless-insight-result-list';
 
 export {buildInsightInteractiveResult} from './controllers/insight/result-list/headless-insight-interactive-result';
+
+export type {InsightInterfaceState} from './features/insight-interface/insight-interface-state';
+
+export type {InsightInterface} from './controllers/insight-interface/insight-interface';
+export {buildInsightInterface} from './controllers/insight-interface/insight-interface';