+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,158 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.15.19](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.18...@coveo/atomic-react@1.15.19) (2022-07-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.18](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.17...@coveo/atomic-react@1.15.18) (2022-07-25)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.17](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.16...@coveo/atomic-react@1.15.17) (2022-07-22)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.16](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.15...@coveo/atomic-react@1.15.16) (2022-07-21)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.15](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.14...@coveo/atomic-react@1.15.15) (2022-07-21)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.14](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.13...@coveo/atomic-react@1.15.14) (2022-07-15)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.13](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.12...@coveo/atomic-react@1.15.13) (2022-07-15)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.11...@coveo/atomic-react@1.15.12) (2022-07-14)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.11](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.10...@coveo/atomic-react@1.15.11) (2022-07-14)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.10](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.9...@coveo/atomic-react@1.15.10) (2022-07-14)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.9](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.8...@coveo/atomic-react@1.15.9) (2022-07-14)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.8](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.7...@coveo/atomic-react@1.15.8) (2022-07-13)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.7](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.6...@coveo/atomic-react@1.15.7) (2022-07-12)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.5...@coveo/atomic-react@1.15.6) (2022-07-11)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.4...@coveo/atomic-react@1.15.5) (2022-07-11)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.3...@coveo/atomic-react@1.15.4) (2022-07-08)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.2...@coveo/atomic-react@1.15.3) (2022-07-08)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.1...@coveo/atomic-react@1.15.2) (2022-07-08)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.15.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.15.0...@coveo/atomic-react@1.15.1) (2022-07-07)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 # [1.15.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.14.1...@coveo/atomic-react@1.15.0) (2022-07-07)