+++ packages/bueno/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.42.1](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.42.0...@coveo/bueno@0.42.1) (2022-07-14)
+
+**Note:** Version bump only for package @coveo/bueno
+
+
+
+
+
 # [0.42.0](https://github.com/coveo/ui-kit/compare/@coveo/bueno@0.41.1...@coveo/bueno@0.42.0) (2022-07-07)